styling upgrade, test structure changes and tests fix
