+++ .vscode/settings.json
@@ -1,3 +1,6 @@
 {
-    "python.pythonPath": "/home/oskar/.virtualenvs/quizy/bin/python"
+    "python.pythonPath": "/home/oskar/.virtualenvs/quizy/bin/python",
+    "python.linting.pylintEnabled": true,
+    "python.linting.enabled": true,
+    "python.linting.pylintArgs": ["--load-plugins", "pylint_django"]
 }
