+++ quizysite/quizyapp/templates/quizyapp/quiz_params.html
@@ -4,13 +4,10 @@
 {% block content %}
 
 
-<div>
-
-    <form role="form" class="form-horizontal" action="/quiz/params/" method="post">
+<form role="form" class="form-horizontal" action="/quiz/params/" method="post">
     {% csrf_token %}
     {% bootstrap_form form layout="horizontal" %}
     {% buttons submit='OK' reset='default' layout='horizontal' %}{% endbuttons %}
-    </form>
+</form>
 
-</div>
 {% endblock %}
