+++ quizysite/templates/base.html
@@ -58,10 +58,11 @@
 
 
     <div class="container">
-
-        {% block content %}
-        {% endblock %}
-    </div>
+        <div class="row">
+            <div class="col-lg-4 col-md-6 col-sm-8 ml-auto mr-auto">
+                {% block content %}
+                {% endblock %}
+            </div>
 </body>
 
 </html>
