+++ quizysite/users/templates/registration/login.html
@@ -1,12 +1,12 @@
 {% extends 'base.html' %}
-
+{% load bootstrap4 %}
 {% block content %}
-<h2>Login</h2>
 
+<h2>Login</h2>
 <form method="post">
     {% csrf_token %}
-    {{ form.as_p }}
-    <input type="submit" value="Login">
+    {% bootstrap_form form %}
+    {% buttons submit='Login' %}{% endbuttons %}
 </form>
 
 {% endblock %}
