+++ quizysite/users/templates/users/register.html
@@ -1,12 +1,13 @@
 {% extends 'base.html' %}
+{% load bootstrap4 %}
 
 {% block content %}
 <h2>Register</h2>
 
 <form method="post">
     {% csrf_token %}
-    {{form}}
-    <input type="submit" value="Register">
+    {% bootstrap_form form %}
+    {% buttons submit='Register' %}{% endbuttons %}
 </form>
 
 {% endblock %}
